algo/knowledge-service/requirements.txt
# FastAPI and dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Neo4j
neo4j==5.16.0

# NLP
spacy==3.7.2
# Note: Download models with:
# python -m spacy download en_core_web_sm  # English
# python -m spacy download zh_core_web_sm  # Chinese

# Redis
redis[hiredis]==5.0.1

# Kafka
confluent-kafka==2.3.0

# HTTP Client
httpx==0.26.0

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
opentelemetry-exporter-otlp==1.22.0
prometheus-client==0.19.0

# Utilities
python-json-logger==2.0.7

# Knowledge Graph Enhancement
networkx==3.2.1
python-louvain==0.16
sentence-transformers==2.3.1

# Object Storage
minio==7.2.3

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
